refactor(Mapper): create classes with extension functions to convert Dtos to Entities and way around
