+++ src/main/kotlin/ecommerce/mapper/CartItemMapper.kt
@@ -1,0 +1,16 @@
+package ecommerce.mapper
+
+import ecommerce.dto.CartItemResponse
+import ecommerce.model.CartItem
+
+class CartItemMapper
+
+fun CartItem.toDto(): CartItemResponse {
+    return CartItemResponse(
+        product.id,
+        productName = product.name,
+        quantity = quantity,
+        productPrice = product.price,
+        productImageUrl = product.imageUrl,
+    )
+}
