+++ src/main/kotlin/ecommerce/mapper/MemberMapper.kt
@@ -1,0 +1,8 @@
+package ecommerce.mapper
+
+import ecommerce.dto.MemberDto
+import ecommerce.model.Member
+
+class MemberMapper
+
+fun Member.toDto() = MemberDto(id, email, role)
