+++ src/main/kotlin/ecommerce/mapper/OptionMapper.kt
@@ -1,0 +1,14 @@
+package ecommerce.mapper
+
+import ecommerce.dto.OptionDto
+import ecommerce.model.Option
+
+class OptionMapper
+
+fun Option.toDto(): OptionDto {
+    return OptionDto(name, quantity)
+}
+
+fun OptionDto.toEntity(): Option {
+    return Option(name, quantity)
+}
