+++ src/main/kotlin/ecommerce/model/Member.kt
@@ -1,6 +1,5 @@
 package ecommerce.model
 
-import ecommerce.dto.MemberDto
 import jakarta.persistence.Entity
 import jakarta.persistence.GeneratedValue
 import jakarta.persistence.GenerationType
@@ -15,8 +14,4 @@ class Member(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     val id: Long = 0L,
-) {
-    fun toDto(): MemberDto {
-        return MemberDto(id, email, role)
-    }
-}
+)
