+++ src/main/kotlin/ecommerce/model/Option.kt
@@ -1,6 +1,5 @@
 package ecommerce.model
 
-import ecommerce.dto.OptionDto
 import ecommerce.exception.InsufficientQuantityException
 import jakarta.persistence.Entity
 import jakarta.persistence.GeneratedValue
@@ -20,10 +19,6 @@ class Option(
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     val id: Long = 0L,
 ) {
-    fun toDto(): OptionDto {
-        return OptionDto(name, quantity)
-    }
-
     fun reduceOptionQuantity(value: Int) {
         if (quantity < value) {
             throw InsufficientQuantityException("Insufficient quantity: $quantity")
